+++ pmd-modelica/src/main/java/net/sourceforge/pmd/lang/modelica/ast/AbstractModelicaClassSpecifierNode.java
@@ -4,7 +4,7 @@
 
 package net.sourceforge.pmd.lang.modelica.ast;
 
-import net.sourceforge.pmd.lang.modelica.resolver.InternalModelicaResolverApi;
+import net.sourceforge.pmd.lang.modelica.resolver.InternalApiBridge;
 import net.sourceforge.pmd.lang.modelica.resolver.ModelicaClassType;
 
 /**
@@ -34,14 +34,14 @@ abstract class AbstractModelicaClassSpecifierNode extends AbstractModelicaNode i
         for (int i = 0; i < listNode.getNumChildren(); ++i) {
             AbstractModelicaNode child = (AbstractModelicaNode) listNode.getChild(i);
             if (child instanceof ASTExtendsClause) {
-                InternalModelicaResolverApi.addExtendToClass(
+                InternalApiBridge.addExtendToClass(
                         classTypeDeclaration,
                         listNode.getVisibility(),
                         child.firstChild(ASTName.class).getCompositeName()
                 );
             }
             if (child instanceof ASTImportClause) {
-                InternalModelicaResolverApi.addImportToClass(
+                InternalApiBridge.addImportToClass(
                         classTypeDeclaration,
                         listNode.getVisibility(),
                         child.firstChild(ModelicaImportClause.class)